Put in place the basic framework of the new Python setup

+++ install.sh
@@ -1,6 +1,9 @@
 #!/usr/bin/env bash
 
 python3 -m venv .venv
+./.venv/bin/pip install --upgrade pip
 ./.venv/bin/pip install -e .
 ./.venv/bin/pip install -r ./dev-requirements.txt
-ln -s ./.venv/bin/blab .
+if [[ ! -e blab ]] ; then
+  ln -s ./.venv/bin/blab .
+fi
